+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Other Changes
 
+- Updated `Microsoft.Identity.Client` and `Microsoft.Identity.Client.Extensions.Msal` dependencies to version 4.80.0.
+
 ## 1.18.0-beta.2 (2025-11-19)
 
 ### Other Changes